[BSK-1749] removed additional build warnings.  Caught an issue where the dvAccumulation output message was in units of nano-seconds instead of seconds.  This has now been fixed and the unit test corrected.
Former-commit-id: 14960296fbf6e44efc3e38621316fe3dbf775699
Former-commit-id: 458844edd6dda954836cc1fd9bb369338406f661

+++ src/fswAlgorithms/attDetermination/_GeneralModuleFiles/ukfUtilities.c
@@ -256,7 +256,7 @@ int32_t ukfLUBckSlv(double *sourceMat, int32_t nRow, int32_t nCol,
 				sum -= sourceMat[i*nRow + j] * destMat[j];
 			}
 		}
-		else if (sum)
+		else if ((int) sum)
 		{
 			ii = i;
 		}
